requirements.txt
PySide6==6.8.2.1
python-dotenv==1.0.1
cryptography==44.0.2 